requirements.txt
numpy
PyOpenGL
PyOpenGL_accelerate
glfw
